Initial lock.h header separation, build and test on Windows
