+++ include/os/windows/etcpal/os_rwlock.h
@@ -17,8 +17,8 @@
  * https://github.com/ETCLabs/EtcPal
  ******************************************************************************/
 
-#ifndef ETCPAL_OS_LOCK_H_
-#define ETCPAL_OS_LOCK_H_
+#ifndef ETCPAL_OS_RWLOCK_H_
+#define ETCPAL_OS_RWLOCK_H_
 
 #ifndef NOMINMAX
 #define NOMINMAX 1    /* Suppress some conflicting definitions in the Windows headers */
@@ -36,38 +36,6 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-/*
- * SRWLocks have been shown by empirical testing to provide a large efficiency boost over Windows
- * mutexes and critical sections.
- */
-typedef struct
-{
-  bool    valid;
-  SRWLOCK lock;
-} etcpal_mutex_t;
-
-#define ETCPAL_MUTEX_HAS_TIMED_LOCK 0
-
-bool etcpal_mutex_create(etcpal_mutex_t* id);
-bool etcpal_mutex_lock(etcpal_mutex_t* id);
-bool etcpal_mutex_try_lock(etcpal_mutex_t* id);
-bool etcpal_mutex_timed_lock(etcpal_mutex_t* id, int timeout_ms);
-void etcpal_mutex_unlock(etcpal_mutex_t* id);
-void etcpal_mutex_destroy(etcpal_mutex_t* id);
-
-typedef HANDLE etcpal_signal_t;
-
-#define ETCPAL_SIGNAL_HAS_TIMED_WAIT 1
-#define ETCPAL_SIGNAL_HAS_POST_FROM_ISR 0
-
-bool etcpal_signal_create(etcpal_signal_t* id);
-bool etcpal_signal_wait(etcpal_signal_t* id);
-bool etcpal_signal_try_wait(etcpal_signal_t* id);
-bool etcpal_signal_timed_wait(etcpal_signal_t* id, int timeout_ms);
-void etcpal_signal_post(etcpal_signal_t* id);
-#define etcpal_signal_post_from_isr etcpal_signal_post
-void etcpal_signal_destroy(etcpal_signal_t* id);
 
 typedef struct
 {
@@ -89,23 +57,8 @@ bool etcpal_rwlock_timed_writelock(etcpal_rwlock_t* id, int timeout_ms);
 void etcpal_rwlock_writeunlock(etcpal_rwlock_t* id);
 void etcpal_rwlock_destroy(etcpal_rwlock_t* id);
 
-typedef HANDLE etcpal_sem_t;
-
-#define ETCPAL_SEM_HAS_TIMED_WAIT 1
-#define ETCPAL_SEM_HAS_POST_FROM_ISR 0
-#define ETCPAL_SEM_HAS_MAX_COUNT 1
-#define ETCPAL_SEM_MUST_BE_BALANCED 0
-
-bool etcpal_sem_create(etcpal_sem_t* id, unsigned int initial_count, unsigned int max_count);
-bool etcpal_sem_wait(etcpal_sem_t* id);
-bool etcpal_sem_try_wait(etcpal_sem_t* id);
-bool etcpal_sem_timed_wait(etcpal_sem_t* id, int timeout_ms);
-bool etcpal_sem_post(etcpal_sem_t* id);
-#define etcpal_sem_post_from_isr etcpal_sem_post
-void etcpal_sem_destroy(etcpal_sem_t* id);
-
 #ifdef __cplusplus
 }
 #endif
 
-#endif /* ETCPAL_OS_LOCK_H_ */
+#endif /* ETCPAL_OS_RWLOCK_H_ */
