+++ src/os/windows/etcpal/os_mutex.c
@@ -17,28 +17,56 @@
  * https://github.com/ETCLabs/EtcPal
  ******************************************************************************/
 
-#include "unity_fixture.h"
+#include "etcpal/mutex.h"
 
-extern "C" void run_all_tests(void)
+bool etcpal_mutex_create(etcpal_mutex_t* id)
 {
-  RUN_TEST_GROUP(etcpal_cpp_error);
-  RUN_TEST_GROUP(etcpal_cpp_uuid);
-#if !ETCPAL_NO_OS_SUPPORT
-  RUN_TEST_GROUP(etcpal_cpp_log_timestamp);
-  RUN_TEST_GROUP(etcpal_cpp_log);
-  RUN_TEST_GROUP(etcpal_cpp_mutex);
-  RUN_TEST_GROUP(etcpal_cpp_rwlock);
-  RUN_TEST_GROUP(etcpal_cpp_sem);
-  RUN_TEST_GROUP(etcpal_cpp_signal);
-  RUN_TEST_GROUP(etcpal_cpp_thread);
-  RUN_TEST_GROUP(etcpal_cpp_timer);
+  if (id)
+  {
+    InitializeSRWLock(&id->lock);
+    id->valid = true;
+    return true;
+  }
+  return false;
+}
 
-#if !DISABLE_QUEUE_TESTS
-  RUN_TEST_GROUP(etcpal_cpp_queue);
-#endif
+bool etcpal_mutex_lock(etcpal_mutex_t* id)
+{
+  if (id && id->valid)
+  {
+    AcquireSRWLockExclusive(&id->lock);
+    return true;
+  }
+  return false;
+}
 
-#endif
-#if !ETCPAL_NO_NETWORKING_SUPPORT
-  RUN_TEST_GROUP(etcpal_cpp_inet);
-#endif
+bool etcpal_mutex_try_lock(etcpal_mutex_t* id)
+{
+  if (id && id->valid)
+  {
+    return TryAcquireSRWLockExclusive(&id->lock);
+  }
+  return false;
+}
+
+bool etcpal_mutex_timed_lock(etcpal_mutex_t* id, int timeout_ms)
+{
+  return (timeout_ms == 0) ? etcpal_mutex_try_lock(id) : etcpal_mutex_lock(id);
+}
+
+void etcpal_mutex_unlock(etcpal_mutex_t* id)
+{
+  if (id && id->valid)
+  {
+    ReleaseSRWLockExclusive(&id->lock);
+  }
+}
+
+void etcpal_mutex_destroy(etcpal_mutex_t* id)
+{
+  if (id && id->valid)
+  {
+    // SRWLocks do not need to be destroyed.
+    id->valid = false;
+  }
 }
