+++ tests/integration/rwlock_integration_test.c
@@ -17,7 +17,7 @@
  * https://github.com/ETCLabs/EtcPal
  ******************************************************************************/
 
-#include "etcpal/lock.h"
+#include "etcpal/rwlock.h"
 #include "unity_fixture.h"
 
 #include <stdio.h>
