+++ tests/unit/live/CMakeLists.txt
@@ -12,7 +12,10 @@ etcpal_add_live_test(etcpal_live_unit_tests C
 
 if(ETCPAL_HAVE_OS_SUPPORT)
   target_sources(etcpal_live_unit_tests PRIVATE
-    test_lock.c
+    test_mutex.c
+    test_rwlock.c
+    test_sem.c
+    test_signal.c
     test_timer.c
     test_thread.c
   )
